requirements.txt
APScheduler==3.10.4
asterisk_ami==0.1.6
panoramisk==1.4
SQLAlchemy==2.0.19
PyMySQL==1.0.3
python-socketio==5.7.0
Flask==2.3.2
Flask-SocketIO==5.3.2
Flask-Cors==3.0.10
eventlet==0.33.3
requests==2.31.0
gunicorn
websocket-client